sistemata cerewazioe appartamento utilizzando dto   sistemato il service per il relativo mapping   bug fix

+++ core-service/src/main/java/com/propertypilot/coreservice/service/TenantServiceImpl.java
@@ -59,7 +59,7 @@ public class TenantServiceImpl implements TenantService {
     @Transactional
     @Override
     public Tenant createTenant(CreateTenantDTO dto) {
-        User user = userRepository.findById(dto.getUserId())
+        User user = userRepository.findById((long) dto.getUserId())
                 .orElseThrow(() -> new RuntimeException("User non trovato"));
 
         if (user.getTenantKey() != null) {
